requirements.txt
pyspark
findspark
boto3
papermill
pyarrow
pandas
langchain
ollama
langchain-ollama
langchain-chroma
weaviate-client
langchain_community
langchain_core
fitz
uuid
pdfplumber
langchain_cohere
rank_bm25
nltk
langgraph
langgraph_bigtool
langchain-groq
oracledb
tabulate
python-dotenv
langchain_google_genai 
google-adk==0.3.0
google-cloud-aiplatform[adk,agent_engines,database]
build
graphviz